refactor: collapse feature-gated items into their owning modules
Apply the review findings from /simplify:

Move Claude-specific helpers out of the shared model module:
  - SessionFile, MAX_FILE_ACCESSES, truncate_string all lived in
    src/model/session.rs behind #[cfg(feature = "claude")] gates,
    even though they are used exclusively by src/collector/claude.rs.
  - Moved them to claude.rs. The serde::Deserialize import is no
    longer needed in model/session.rs either.
  - Net: 4 cfg annotations removed, one cross-module dependency gone.

Consolidate the Codex rate-limit code in rate_limit.rs:
  - 5 items (one const, two pub fns, two private fns) each carried
    their own #[cfg(feature = "codex")] attribute.
  - Collapsed into a single cfg-gated mod codex_cache { ... } plus a
    cfg-gated pub use re-export, preserving the existing public path
    (rate_limit::read_codex_cache, rate_limit::write_codex_cache).

No behavior change. cargo test passes 73/73, clippy clean on default,
claude-only, and codex-only builds.
